login: Validate the user's organization on login
In order to use any Upbound features, the user needs to have an
organization. The organization is stored, somewhat confusingly, in the user's up
profile as the "account", and can be provided using the `-a` option at
login. Previously, `up` was happy to accept any value for the organization name:
it could be an organization that didn't exist, or one the user wasn't part of.

Add two validation steps during `up login` to ensure the user has a valid
organization:

1. If the user didn't provide an organization name (with the flag or in their
   existing up profile), check whether they have any organizations. If they
   don't have any, throw an error. If they have at least one, use the first one.
2. After the user is logged in, try to get the organization ID of the
   organization in their up profile. This validates that the organization exists
   and they have access to it.

Fixes #255

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>