project: Automatically move projects during `up project run`
If we detect that a user is trying to run a project whose repository isn't owned
by their organization, automatically move it (without writing any changes to
disk) rather than telling the user to move it manually. This maximizes the
chance that `up project run` will Just Work when a user checks out an example
repository.

Keep the existing checks if the user specifies a repository explicitly, since it
would be surprising for `up project run` to use a different repository than
specified on the commandline, but still move the project if they've specified a
repository they own.

Fixes #257

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>