project: Pass the real base path in `up project run`
When we build functions, we need their real base path in the OS filesystem. We
can infer it when the function's FS is an afero.BasePathFs, but if it's a
CopyOnWriteFs as in `up project run` we get the wrong path and can't follow
symlinks.

Pass the path explicitly in `up project run`. In the future we should probably
always pass it so we're not making assumptions about the filesystem types in the
builder.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>