profile: Rename "account" to "organization" in config files
The "account" field in a profile is actually the organization on which to
operate. Add an `Organization` field to the profile, and use it instead of the
`Account` field when constructing contexts. Construct and update profiles with
`Organziation` filled in instead of `Account`, so that existing configurations
will automatically be migrated.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>