space: Create a disconnected profile in `up space init`
When initializing a new space, create a profile for it if the user has not
already selected a disconnected space with a kubeconfig. Copy the current
profile's Upbound login credentials to the new profile if they exist, so the new
profile can use Marketplace. Activate the new profile so the user is ready to
create control planes in the new space.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>