ctx: Rename the kubeconfig flag
`up ctx` has a non-standard short form for `--kubeconfig`, `-f`, which we'd like
to retain. It also accepts `-` as an argument to allow writing the kubeconfig to
stdout.

Since `--kubeconfig` is now part of the standard upbound flags, we need to
rename `ctx`'s special version of it. Make it `--file` and treat it as an
override for `--kubeconfig`.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>