ctx: Make the top level profile-dependent
When a cloud profile is active, use its organization as the top level in `up
ctx`. When a disconnected profile is active, make its disconnected space the top
level. This makes `up ctx` profile-centric. If the current context doesn't match
the current profile, revert to the top level for the profile.

Fixes #300

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>