ctx: Simplify kubeconfig merging and make results more intuitive
Previously, the logic used to merge a new context into the existing kubeconfig
was rather hard to understand. Simplify, and make the results more intuitive in
some cases (as shown by the TODO removal in the tests).

Critically, the merge logic no longer uses `activateContext`, so it doesn't need
to move when we factor out the kubeconfig merging from ctx.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>