projects: Rename `up project run`'s --allow-production flag
The `--allow-production` flag tells `up project run` to use an arbitrary control
plane, skipping the check that ensures it only uses dev MCPs. This is necessary
for self-hosted Spaces users, since dev MCPs are cloud-only. However, the name
can be confusing since the control planes these users have are not necessarily
"production".

Rename the flag to `--skip-control-plane-check`, keeping `--allow-production` as
an alias in case anyone's using it already. The new name makes it more clear
what the flag does, retains a vague sense of danger, and is harder to set
unintentionally than something generic like `--force`.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>