profile: Handle field upgrades and defaults when unmarshaling
When we populate an `*upbound.Context` based on the user's configuration, we
handle upgrading the old `Account` field to `Organization` and treat profiles
with an empty `Type` field as cloud profiles. However, we weren't
upgrading/defaulting these in the raw profiles, so commands that fetched the
active profile and used these fields would get incorrect data.

Implement a custom unmarshaler that sets `Organization` to the value of
`Account` when necessary and sets a default for `Type`. In addition to fixing
commands that previously assumed `Type` and `Organization` would be set, this
lets us simplify some code that handled them being unset.

Fixes #405

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>