ctx: Explicitly require that a profile exists
Previously, `up ctx` worked for disconnected spaces without an up profile being
created. Now we require a disconnected profile to be created for use with a
disconnected space. Check that the user has a profile and return a clear error
message if they don't, rather than falling through to a later stage and
returning `unknown profile type`.

Fixes #384

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>