projects: Wait for the latest version to be installed in `up project run`
Previously, we were looking for *any* version of the configuration in the
package manager lock when waiting for a project to become ready. This meant that
if a previous version installed successfully, but the latest version couldn't be
installed (e.g., because of auth failures to the registry) we would silently
move on with checking the dependencies and return success despite the update not
being installed.

Look up the current version in the lock to make sure it's been installed. For
now, this will make `up project run` hang forever if the package can't be
installed.

Fixes #397

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>