ctx: Improve error message when profile is set incorrectly
When using the non-interactive mode of `up ctx`, it's possible to request a
context for a different profile than the current one. Previously this would
result in a confusing error message like `"upbound" not found in: awg-test`.

Check that the requested context has the right prefix for the current profile
before trying to navigate to it, and return a clear error message if it doesn't.

Fixes #407

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>