projects: Add support for building Go embedded functions
Detect that a function is Go by looking for the `go.mod` file. Build Go
functions using `ko` as a library. Use the same base images that our ecosystem
team uses to build official providers.

`ko` pulls in a newer version of go-containerregistry, which requires a small
change in our unit tests. That change is included in this commit.

Fixes #454

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>