projects: Fix moving projects on in-memory filesystems
In `up project build` and `up project run` we "move" projects to a different
repository in an in-memory filesystem before building, so that the function
references in compositions match the project's repository location. We use
afero's `CopyOnWriteFs` for this, creating an in-memory overlay filesystem. The
`CopyOnWriteFs` treats relative paths differently than the `BasePathFs` we
usually use, which meant our project move wasn't effective when using the
in-memory filesystem.

Create a no-op `BasePathFs` on top of the `CopyOnWriteFs` so that relative paths
work as expected.

Fixes #444

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>