crd: Fill in list kinds if Crossplane doesn't produce them
When Crossplane generates a CRD, it doesn't fill in the list kind. This results
in wonky OpenAPI specs being generated by our OpenAPI conversion, since the list
kind is generated but doesn't have a name. Fill in the list kind by appending
`List` to the kind name so we end up with nice OpenAPI specs.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>