profile: Use new profiles by default
When a profile is created via `up login --profile` or `up space init`, it
becomes the current profile automatically. This is not the case for profiles
created with `up profile create`, which is confusing.

Change the default for the `--use` flag to true so new profiles become current
by default.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>