async: Avoid creating any spinner infrastructure in quiet mode
async.WrapWithSuccessSpinners creates a writer to display its spinners,
regardless of whether any spinners end up being displayed. This outputs some
control characters that are non-visible but still present up if the user is
piping output somewhere.

Introduce a wrapper function type in the async package that
WrapWithSuccessSpinners conforms to and offer an alternative, IgnoreEvents, that
callers can use to run the same code but not write anything. In code that uses
the async spinners, select an appropriate wrapper function depending on the
quiet flag.

Fixes #513

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>