Factor out registry keychain creation
We have a bunch of commands that create registry keychains for authentication
when working with the xpkg registry. Usually we want a keychain that respects
the user's local docker config and also uses our credential helper to
authenticate to xpkg for the current context's organization. Factor out creation
of this kind of keychain so it's easy to reuse.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>