*: Remove short flag for cache directory
`-d` is the global flag for debug output, defined in our shared
`upbound.Flags`. Not all commands currently take these flags, but they
should. Some commands had redefined `-d` as the short flag for the package cache
directory; this option is likely not used much, and the short alias conflicts
when we add the global flags.

Remove all the `-d` short aliases from cache directory flags to solve the
conflict.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>