filesystem: Swich tests from testify/require to gotest.tools/assert
I noticed renovate was trying to upgrade testify/require and wondered why we
have a dependency on it since our linter forbids its use. We haven't formally
blessed gotest.tools/assert in the style guide, but we are using it widely.

Switch the filesystem tests from testify/require to gotest.tools/assert so we no
longer have a direct dependency on testify/require.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>