projects: Add support for building go-templating functions
Add a function builder for go-templating functions. We accept any directory
containing only `.tmpl` and/or `.gotmpl` files. To build the function we simply
add the source directory to the base image and set the environment variable
`FUNCTION_GO_TEMPLATING_DEFAULT_SOURCE` so the function will load the templates
at runtime.

Fixes #578

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>