projects: Remove dependency on the go binary for schema generation
By default, oapi-codegen formats the code it produces using
`golang.org/x/tools/imports`, which requires the go binary be available in the
user's PATH. Many of our users will not have go installed, so schema generation
won't work for them and since `up project build` always generates schemas, they
generally won't be able to use projects.

Set the `SkipFmt` option on our oapi-codegen call so that it doesn't attempt to
format the code, thus removing the dependency on the go binary. Format the code
ourselves instead using the `go/format` package, which is
self-sufficient. Because we're not using the imports formatter, this does mean
we end up with some unused imports in our schemas, which are harmless other than
making the linter mad about the generated code.

Fixes #617

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>