profile: Serialize versioned kubeconfigs for disconnected profiles
In memory, we need a `clientcmdapi.Config` to connect to Spaces. But, that's not
meant for serialization to disk; for that we should use versioned
types. Previously, we were serializing the unversioned config, which can't be
deserialized if it contains extensions. This led to users creating broken
configs when they accidentally created profiles while pointed at an up-created
kubeconfig context.

Change our profile marshaling to convert the in-memory kubeconfig to a versioned
one before writing to disk. When we unmarshal, detect which version was written
and populate the unversioned in-memory kubeconfig appropriately. This has the
side-effect of automatically upgrading existing up configurations to use the new
format.

Fixes #568