projects: Ensure image digests don't change when we write them to disk
go-containerregistry's "tarball" format, which we use for writing packages to
disk, does not preserve the original manifest when it writes an image; rather,
it reconstructs it from its own manifest format. The reconstructed manifest
always uses the docker media types, so if the original image used the OCI media
types the digest of the image will change.

Round-trip embedded function images through the tarball format to ensure their
digests won't change when we write them out in tarball format later. This
ensures the dependency digests we generate for configurations are correct.

Fixes #639

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>