projects: Avoid round-tripping images through tarball during build
Previously, we introduced an in-memory round-trip through go-containerregistry's
tarball format to ensure image digests wouldn't change when written out to disk
by build and read back in for push. This was super slow because (a) we weren't
using the layer cache, and (b) it involves shuffling a lot of bytes.

Using the layer cache for the round-trip would improve performance a bit, but
still involves shuffling a lot of bytes around.

Replace the tarball round-trip with a mutation of the images to use the Docker
media types that are used by the tarball format. This has the same effect as the
round-trip but avoids the need to read all the layers into memory during the
build.

Fixes #704

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>