projects: Fix race condition in Go schema generation
oapi-codegen has a bunch of internal global variables, which is why we already
prevent concurrent calls to `codegen.Generate`. It turns out one of these
globals is also used in the codegen pacakge's name generation functions, which
we were calling concurrently.

Acquire the Go code generation mutex when generating names to avoid the race.

Fixes #711

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>