projects: Ensure Python files are readable in function images
Crossplane runs function pods with UID 2000 and GID 2000. If files in an
embedded Python function are not world-readable and the local user's UID isn't
2000, then the interpreter function isn't able to read them in the resulting
function container and it fails to run.

Make the Python files owned by UID 2000 and GID 2000 so that the interpreter
function will be able to read them.

Fixes #700

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>