crd: Generate enums for APIVersion and Kind
We already generate default values for APIVersion and Kind when creating OpenAPI
schemas for CRDs. Generate enum values for them as well, so that each model has
only a single valid value for each. This makes each resource unambiguously match
a single schema when using JSON Schemas with YAML documents, once the APIVersion
and Kind are filled in.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>