ctp: Add support for local dev control planes
Refactor internal/ctp to support creating two different kinds of control planes:
local with kind, or Spaces. Spaces dev control planes are the default when the
user's current context is in a Space; local are the default otherwise. A
`--local` flag lets users force a local control plane regardless of their
current context.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>