ctp: Sideload packages into local dev control planes
Sideload packages by writing them into an OCI layout and running olareg as a
registry with the layout as its storage backend. This avoids the need to
separately load the images into kind/containerd and Crossplane, which don't
share a cache, and is faster than pushing to a local registry.

We sideload packages using their original repository paths by creating an
ImageConfig in the cluster that rewrites the original repository path to the
local dev registry. This lets us build packages with the original repository
regardless of what kind of runner we're using.

With this change, users no longer need to be logged into Upbound to use `up
project run`.

Fixes #882

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>