ctp: Create xpkg pull secrets and ImageConfig to use them
If a user is logged in when we create a local dev control plane, create a pull
secret using their session credentials and configure Crossplane to use it via
ImageConfig. This gives local dev control planes access to private repositories,
which are the default for `up project run` and `up test`.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>