xpkg: Remove build-time schema generation
UXPv2 will have support for composing arbitrary resources. In order to support
that in our DevEx, we need to generate schemas for arbitrary resources, and the
best way to do that is to generate schemas on-the-fly on the client side.

As a first step, remove all the code for generating and packaging schemas as
part of package builds. This includes the schema-generator tool used in our
official package build pipeline, which will no longer be necessary.

Upcoming commits will move code around and then reintroduce schema building to
the DevEx commands.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>