projects: Stop parsing workspaces where possible
Parsing a workspace involves reading and parsing every single yaml file in the
directory. In many places, we were parsing a workspace just to read and update
the `upbound.yaml` file. Add some new helper functions to the project package
for updating project metadata and use them instead of workspaces where
possible. This should be more efficient and make the code easier to understand.

The one remaining place where project commands use workspaces is in
project.Move. Here, we need to update compositions as well as metadata, so it
makes sense to lean on the workspace for finding and parsing all the
compositions in the directory.

Since we're updating the spots where we deal with dependencies, also take the
opportunity to move to new-style APIVersion/Kind dependencies.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>