kube: Take index.docker.io into account when looking up packages
For bad historical reasons, OCI references that start with docker.io are often
rewritten to start with index.docker.io. The go-containerregistry package does
this when parsing repository names.

If someone uses docker.io for their project's OCI repository, `up project run`
will be looking for index.docker.io in the package manager lock, but will find
docker.io, since Crossplane doesn't rewrite the reference for lock
purposes. This caused our package lookup logic to fail to find the package and
spin forever waiting for it to exist.

Parse the references so that this (and any other) defaulting will be applied
before comparing them.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>