kube: Don't return early when retrying with backoff
Previously, we were using `wait.ExponentialBackoff` from Kubernetes to retry
package installation and wait for packages to become ready. It returns as soon
as the number of `Steps` or the `Cap` is reached, which isn't what we intended:
we set timeouts on our contexts, and should wait until they get cancelled.

Implement our own retry function that returns only on success or when the
context is cancelled. Use the Kubernetes backoff type to calculate retry
intervals so we don't need to implement that ourselves.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>