Re-introduce the schema-generator
Rebase the deleted schema-generator on top of the schema manager refactoring we
did while it was gone. We don't actually use the schema manager itself since we
don't need the lockfile capability and want a filesystem per language, but we do
take advantage of the Source abstraction for extracting resources from
packages. The schemas produced should operate identically to before.

Fixes #955

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>