helm: Handle OCI repositories transparently
Upstream helm now has perfectly good support for OCI registries, so we don't
need to implement it ourselves. However, there's a gotcha: to pull from an OCI
registry, Helm wants the full OCI path to the chart (including `oci://` prefix)
rather than a repo URL and chart name.

Handle constructing the necessary "chart ref" (`oci://registry/repo/chart`) when
we construct the helm manager, so that clients always pass the same parameters
(repo and chart name) regardless of whether they're using OCI or not and the
rest of the helm manager code works the same as well. Clients now pass an
`oci://` repository or one with no scheme to use OCI, and `https://` to use
non-OCI.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>