xpls: Remove support for linting patch and transform compositions
P&T support has been deprecated for a while, and is removed entirely in
Crossplane v2. Remove support for it in xpls in preparation for updating our
Crossplane dependencies to v2.

This also lets us delete a lot of "vendored" internal crossplane code, which was
only used for running P&T compositions during linting.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>