kube: Handle no-op package updates
Previously, if we installed a new version of a configuration and it was
identical to the previously installed version, we would wait forever for it to
become healthy. This is because the revision name doesn't change, and therefore
the package reference doesn't get updated in the lock; we were looking for the
updated image reference in the lock and it never showed up.

Replace our lock lookup with a fetch of the current revision and check whether
it has the right image reference. This will always get updated when the
configuration is updated, even if the revision doesn't change.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>