ctp: Make relevant commands work with non-Spaces control planes
There are three sets of commands that all rightfully belong in the `up
controlplane` tree, but have different operating requirements:

1. Commands that operate on Spaces control planes, and thus require the user to
   be in a Spaces context: create, delete, list, get.
2. Commands that work only within a Spaces control plane: oidc-auth.
3. Commands that work in any kind of control plane: configuration, function,
   provider, pull-secret, connector.

Previously, `up controlplane` required a Spaces context for all commands, but
checked that it was a control plane context for group 2 and 3 commands. Group 1
commands failed if the user's current context was a control plane, even if that
control plane was in a Space.

Update the checks in `up controlplane` to let the commands in group 3 work on
any control plane. This makes them useful for non-Spaces UXP users. Also make
commands in group 1 work when the user's context is a Spaces control plane; they
will operate on the control plane's group in this case. Group 2 still only works
on Spaces control planes.

Fixes #948
Toward #298

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>