ctp: Handle out-of-date and dead registry containers
The local runner runs a registry container to side-load packages into the
control plane. This registry shares a TLS CA with the control plane. If a local
control plane's kind cluster gets torn down, but the registry container remains
running, the next time the local runner is started a new CA certificate will be
generated, but the registry won't be updated to use it. This results in the
control plane being unable to pull the sideloaded packages. Similarly, if the
registry container exists but is dead, the control plane would fail to pull from
it.

When we find and existing registry container, check that the CA certificate is
correct, and re-create it if not. Also try to start the container (which is
idempotent) to ensure it's running.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>