ctp: Check UXP licenses when using arbitrary kubeconfig contexts
Since we've introduced the `--use-current-context` flag to `up project run` and
`up test run --e2e`, users can point these commands at an OSS Crossplane
cluster. We want our DevEx to be exclusive to UXP, so check for a license when
using this flag to ensure the cluster is a UXP cluster.

Note that since we allow community edition UXP clusters, it's fine for the
license to be missing a `secretRef`, and in this case there's no signed license
data to validate. This means users could technically work around the restriction
by manually installing UXP's license CRD into their OSS cluster and creating a
license CR that looks like a community edition license. If they want to go
through that effort, we'll let them.

Fixes #950

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>