renovate: Undo all the ssh hacks we've tried
I'm giving up on making renovate's post-upgrade commands able to read private
repos. Any dep updates that affect generated code will have CI fail and need to
be updated manually, but that's fine for now. Maybe we can figure out how to
make this better in the future (maybe as part of removing the build submodule!).

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>