projects: Use v2alpha1 internally
To add DevEx support for Operations, we'll introduce an additional path
configuration in the project metadata: the directory that holds operations. This
will be present only in v2alpha1 of projects, not v1alpha1, since only
Crossplane v2 supports operations.

This means we'll need to have the v2alpha1 representation available when we
build projects (and in a few other parts of the code), so we can find the
operations. Today, we pass the v1alpha1 representation around everywhere,
converting from v2alpha1 when we parse a project if necessary.

Prepare for adding operations by changing all our internal code to work with the
v2alpha1 representation. We still support v1alpha1 on disk, but the operations
path will be lost when converting to v1alpha1 since it doesn't exist in that
version.

Toward #869

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>