projects: Support generating operation functions in `up function generate`
There's no difference between operation functions and composition functions in
terms of what we template, and all functions will live in the same place in the
project (`functions/`). So, the only change in `up function generate` is that we
need to support adding steps to operation pipelines as well as composition
pipelines.

Abstract the pipeline update a bit so the main command code doesn't need to know
the details. Support all three kinds of operations (oneshot, cron, and
watch). Remove the default function-dummy step generated by `up operation
generate` if it's the only step in the existing pipeline, since users probably
won't actually want it.

Fixes #869

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>