projects: Add capabilities to embedded functions during build
In Crossplane v2, operation functions must have the "Operation" capability in
their metadata. Theoretically, we could figure out which of our embedded
functions are being used for operations and which are being used for composition
and set capabilities appropriately. For now, though, give all functions both
capabilities since there's little risk in doing so and it's simple.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>