uxp: Remove license option from `up uxp install`
The UXP helm chart no longer supports setting a license; licenses must be
applied separately. For now, remove the `--license-file` flag from `up uxp
install` and instruct users to apply a license with `up uxp license apply` if
needed. We may re-introduce the license flag in the future if we get signal that
it would be desired.

Fixes #1108

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>