test: Clean up control planes synchronously when e2e tests pass
In the e2e test runner we create a dev control plane and have a signal handler
to clean it up if a test run gets killed. Previously, we were also handling
normal teardown in the signal handler's goroutine by sending a signal on a
channel to trigger cleanup.

However, since the goroutine runs async and we weren't waiting for it to finish
before returning from the function, it was possible for `up` to exit before the
cleanup happened.

Clean up synchronously in a `defer` rather than in the goroutine to ensure that
cleanup always happens before `up test run --e2e` exits.

Fixes #1085

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>