main: Construct contexts centrally
Allow commands to indicate that they require an upCtx by embedding a new
`upbound.RequiresContext` struct. The main package will then construct a context
from the embedded flags and bind it to kong context for commands to use.

This lets us centrally collect telemetry on whether users are logged in.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>