projects: Make `up project init` and `up project run` work without auth
Users should be able to get started with Upbound projects without logging in or
having an Upbound account. Previously, if the user wasn't logged in `up project
init` and `up project run` would both use an invalid repository path, preventing
the project from building and running.

In `up project init`, use `xpkg.upbound.io/example/...` as the repository if the
user isn't logged in. `example` is a reserved org name in Upbound cloud, so we
won't be stepping on anyone's toes, and it should be fairly obvious that it
needs to be replaced if the user does log in later.

In `up project run`, use `up-project.local/...` as the repository if the user
isn't logged in. We rewrite this to the sideloading registry in the local runner
anyway, and it makes it clear that we're not pushing/pulling the project to/from
Upbound.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>