ctp: Create a pull secret and ImageConfig for current context runners
When using `up project run --use-current-context` with a private repository, the
cluster needs a pull secret and ImageConfig to pull the project's
packages. Previously, the user was on their own to create this and if they
didn't do it then `up project run` would time out since packages would never
become ready.

If the user is logged in, create a pull secret and ImageConfig for them. This
way, as long as they're using our registry `up project run` (and `up test run`)
should Just Work.

Fixes #1049

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>