help: Format help text using markdown
Use the glamour library to render detailed help text using markdown. This is
simpler than the previous custom styling implementation, and gives us a ton of
flexibility in what we print and how we format it.

Define custom light and dark stylesheets based on glamour's defaults, using
Upbound brand colors for common elements.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>