uxp: Use ingress if available for `up uxp web-ui open`
If the user is pointed at a local dev cluster with ingress enabled, use the
ingress to open the Web UI instead of port forwarding.

The way this is implemented, we can potentially extend it to cloud control
planes in the future when they're running UXPv2 (opening them in the Upbound
console).

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>