project: Validate project names in `up project init`
The positional argument to `up project init` is the project name, which also
becomes the directory name by default. It's tempting to specify an absolute
directory path, but that actually has to be done with the `--directory` flag.

Validate that the name is a valid k8s resource name (which should also make it a
valid OCI repo name), so that if a user specifies a path they get an error
instead of a broken project.

Fixes #1131

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>