ctp: Check session expiry if the user is logged in and update secrets
If the user is logged in when they create a local dev control plane, we use
their session token to create a pull secret in the cluster that lets them pull
any private dependencies. Previously, we didn't check whether the session token
was expired, so the user could end up with their control plane unable to pull.

Check the session token before proceeding with control plane creation and return
an error if it's expired. This lets us fail fast rather than timing out waiting
for packages to become ready. We skip the check if the user's profile doesn't
have an organization, since `up project run` should work for users who are not
logged in at all.

Use Patch instead of Create when creating the pull secret, so that we'll update
the secret if the user has renewed an expired session.

Fixes #1196

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>