uxp: Use OCI for both stable and unstable, filtering versions
We're going to keep using the OCI helm repository, since it's more likely that
users in network restricted environments will have access to it than to
charts.upbound.io. However, both pre-release and stable versions go to this
repository, and UXP versions are all "pre-release" in semver terms, so we can't
rely on helm to select the latest version for us.

Add an option in the helm installer to filter versions, and implement version
listing with the filter. This code is somewhat terrible, because helm makes this
a giant pain; maybe we can improve it in the future (or stop supporting non-OCI
repositories entirely 🤞). In any case, the regexp we're using matches what
Spaces uses to select UXP releases, so we're at least consistent.

Fixes #1113

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>