e2e: Update main.k to match reality and pin dependency version
e2e tests started failing when provider-kubernetes v1.0.0 was released, because
it has an additional API group (for namespaced MRs) and their models get added
automatically to the main.k in `up function generate`. Update the expected
main.k to fix the assertion, and pin provider-kubernetes to v1.0.0 in the e2e
test so we don't get any unexpected changes in the future.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>