ci: Selectively upload artifacts and run e2e jobs on blacksmith runners
Previously we ran our e2e jobs on GH runners because downloading the artifacts
from GH was slow on blacksmith runners. However, the GH runners have become very
flaky, so we'd like to use blacksmith everywhere.

Be more selective in what artifacts we upload so that there's less to download,
and switch the e2e jobs to use blacksmith.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>