project: Serialize updates to the project file when adding dependencies
The `DependencyManager.AddAll` method adds passed dependencies
concurrently. This can result in concurrent updates to the in-memory project and
on-disk project file. For the file, each update is a read-modify-write. If the
read phase of one update happens concurrently with the write phase of another
update, the read can read an empty or partial file, resulting in a
failure. We've seen this manifest as a failure to parse the project due to an
unknown apiVersion.

Add a mutex to serialize updates to the in-memory project and on-disk file. The
expensive parts of `Add` (downloading packages, generating schemas, etc.) are
still outside the mutex, so the concurrency is still helpful.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>