ci: Split our CI workflow into reusable workflows
Our CI workflow is too long and has a bunch of conditional steps since it's
triggered by many different things. As a first step toward splitting it into
multiple workflows with more specific purposes, refactor it into a set of
reusable workflows that will be useful in different contexts.

While we're here, remove the detect-noop job. Our CI is pretty fast, and
detect-noop hardly ever skips a run anyway, so running it is overall a waste of
time.