config: Remove the profile "base config" functionality
The "base config" functionality let users override arbitrary flags by default
via a map in the config file. We removed the command for users to set these
configurations a while ago, and no one has complained, so I'm fairly certain no
one was using this. It's super complex, likely unused, and easily confused with
the "base configuration" functionality that we *do* use, so remove it.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>