config: Simplify the base configuration API
Base configuration is a set of key-value configuration options that live at the
top level of the config file (i.e., non-profile-specific). Today, we only use
this today to configure telemetry.

Remove error returns that are always nil, and add a helper method to get
values. This simplifies the telemetry configuration code.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>