telemetry: Report on CI environments
Check the `CI` environment variable, which is commonly set by CI systems, and
include a new telemetry attribute (`ci.env`) if it's set. Also check for the
`GITHUB_ACTIONS` envvar, which is set to `"true"` by GHA, and include a
`ci.provider` attribute if we find it. We can detect more providers in the
future.

Fixes PEX-64

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>