ci: Use GH artifacts to save builds between jobs
Previously, we switched to using Blacksmith's cache to save the `up` binary
between the build job and the e2e jobs. This is faster than GH artifacts, since
the cache is guaranteed to be located near the runner. However, because it's
intended to be a cache, we've seen a few annoying failure cases:

- Saving fails silently, resulting in a failure to restore the cache in the e2e
  job.
- If CI is busy and our cache is pretty full, the build may be evicted before we
  get to use it.
- Restoring fails silently, resulting in an e2e failure.

In the first two cases, the build is gone and we have to re-run the full CI
workflow, which ends up burning more time than we would have taken uploading and
downloading from GH artifacts.

Note that we upload only the `up` binary for the runner's architecture, not the
full `_output` directory. This keeps our artifact quite small (under 100MB), so
the upload/download shouldn't be *too* slow even if we get unlucky with regions.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>